chore: refactored reducer
