+++ src/pages/ProductList/reducer/actionTypes.ts
@@ -1,0 +1,12 @@
+export enum FilterActionType {
+  CLEAR,
+  RATING,
+  PRICE,
+  CATEGORYAI,
+  CATEGORYDISCRETEMATHS,
+  CATEGORYSYSTEMDESIGN,
+  CATEGORYDSA,
+  CATEGORYDBMS,
+  CATEGORYWEBDEV,
+  SORTBYPRICE,
+}
